+++ nano-orm.js
@@ -11,20 +11,41 @@
 "use strict";
 
 let Q = require('q');
+let moment = require('moment');
 
-//////////////////////////////////////////////////////////////////////////////
-/// \brief Creates a new Model
-/// \param table_name   The name of the database table this model wraps
-/// \param model_fields Array of field descriptors. Order must be the same as
-/// the order of columns in the database. Each field descriptor may be either a
-/// strings representing the field name, or objects of the form:
-/// { name     : <string>,
-///   required : <boolean>,
-///   // any options valid in a JSON schema
-/// }
-/// \param options JSON object representing additional options for the model
-///        - id_field -> Name of the id field, defaults to 'id'
-//////////////////////////////////////////////////////////////////////////////
+/**
+ * Creates a new class for representing a Model which may be loaded from
+ * and persisted to the database
+ *
+ * @param {string} table_name - The name of the table in the database the
+ * created Model should wrap
+ *
+ * @param {Object[]} model_fields - Array of field descriptors. Order must be
+ * that of the columns in the database. Each field descriptor may be either a
+ * string holding the field's name, or a JSON schema for the type stored by the
+ * field. Important properties of the JSON schema are outlined below.
+ * @param {string} model_fields[].name - Name of the field. Must be equal to the
+ * name of the column in the database
+ * @param {bool} [model_fields[].required] - Whether the field is required
+ * (IE: cannot be NULL). Defaults to true.
+ * @param {string} [model_fields[].type] - Name of the datatype used to
+ * represent the field. May be one of:
+ * - string
+ * - number
+ * - boolean
+ * - datetime
+ *
+ * @param {string} [model_fields[].format] -
+ *
+ * @param {Object} [options] - JSON object representing the additional options for
+ * the model.
+ * @param {string} [options.id] - Name of the id field, defaults to 'id'
+ *
+ * @todo :TODO: Is there anything other than id field name to put in options?
+ * Should we just have it as a string
+ *
+ * @return Constructor for the created Model class
+ */
 function defineModel(table_name, model_fields, options){
 	//console.info("Creating Model for table: '" + table_name + "'");
 
@@ -48,13 +69,19 @@ function defineModel(table_name, model_fields, options){
 	let id_field_name = "id";
 	if(options && options.id_field){ id_field_name = options.id_field; }
 
-	/////////////////////////////////////////////////////////////////////
-	/// \brief Constructor which creates new instance of some model
-	/// \param field_values Object representing the values to use for the
-	/// fields of the model. Any fields that are not specified will be assigned
-	/// null. If extra values are put in field_values for which a field does
-	/// not actually exist, then this method will throw.
-	/////////////////////////////////////////////////////////////////////
+	let field_mappers = _generateMappingFunctions(model_fields);
+
+	/**
+	 * Creates a new instance of the Model with specified values for fields.
+	 * This function will NEVER affect the database itself.
+	 *
+	 * @constructor
+	 *
+	 * @param {Object} [field_values] JSON object with values to use for the
+	 * fields of the created model instance. Any field's whose value is not
+	 * specified will be set to null. Attempting to set a field which does not
+	 * exist will result in an exception.
+	 */
 	let Model = function(field_values){
 		// Define fields of the model
 		this._fields = {};
@@ -72,7 +99,11 @@ function defineModel(table_name, model_fields, options){
 				if(this._fields[f] === undefined){
 					throw "Attempted to specify value for non-existent field: " + f;
 				}
-				this._fields[f] = field_values[f];
+				if(typeof field_values[f] === 'string'){
+					this._fields[f] = field_mappers[f].fromDb(field_values[f]);
+				} else {
+					this._fields[f] = field_values[f];
+				}
 			}
 		}
 
@@ -105,27 +136,84 @@ function defineModel(table_name, model_fields, options){
 		}
 	};
 
+	Model._mappers = field_mappers;
+
 	Model._queries = {}; // cache of sql statements
+
+	/**
+	 * Retrieves the name of the table this Model wraps
+	 * @return {string} Name of wrapped table
+	 */
 	Model.getTableName   = (function(){ return table_name;                      });
+
+	/**
+	 * Retrieves list of field names in the Model; NOTE: this does not include the
+	 * name of the id field
+	 * @see {@link Model.getIdFieldName}
+	 *
+	 * @return {string[]} Array of field names in this model
+	 */
 	Model.getFieldNames  = (function(){ return model_fields.map((f) => f.name); });
+
+	/**
+	 * Retrieves the name of the id field for this model.
+	 *
+	 * instance.id can always be used to access the value of the id - this
+	 * function is useful only if writing plain SQL queries that require the id
+	 * field name
+	 *
+	 * @return {string} Name of id field
+	 */
 	Model.getIdFieldName = (function(){ return id_field_name;                   });
 
-	Model.prototype.toJSON = function(){ return this._fields; };
+	/**
+	 * Converts a Model to an object that which should be converted to JSON
+	 * to represent the Model's data. This will include the values of the
+	 * Model's fields only, with no associated meta data (eg: the name of
+	 * table which stores Model
+	 *
+	 * The returned object can be turned back into a Model using the createFromRow
+	 * function.
+	 *
+	 * @return {object}
+	 */
+	Model.prototype.toJSON = function(){
+		let result = {};
+		for(let key of Object.keys(this._fields)){
+			result[key] = Model._mappers[key].toDb(this._fields[key]);
+		}
+		return result;
+	};
 
-	/////////////////////////////////////////////////////////////////////
-	/// \brief Creates a new instance of Model and persists it to the database
-	/// \return Promise which resolves to the model instance after it has
-	/// been saved
-	/////////////////////////////////////////////////////////////////////
+	/**
+	 * Creates a new instance of the Model and persists it to the database
+	 *
+	 * @param dbh {DbConnectionPromise} -  database handle which which to save
+	 * the Model to the database
+	 * @param field_values {object} - Object containing values for the fields
+	 * of the created instance
+	 *
+	 * @return {Promise} Promise which will either resolve to the created
+	 * instance after it has been saved (and thus assigned an id if the id
+	 * column is AUTO_INCREMENT) or be rejected if an error occurs while saving
+	 */
 	Model.create = function(dbh, field_values){
 		let result = new Model(field_values);
 		return result.save(dbh);
 	};
 
-	/////////////////////////////////////////////////////////////////////
-	/// \brief Creates an instance of this model from a row loaded from the
-	/// database are returned by db-connection-promise.query
-	/////////////////////////////////////////////////////////////////////
+	/**
+	 * Creates a new instance of Model from a row of data retrieved from the
+	 * database
+	 *
+	 * In general users of nano-orm should prefer methods such as find() or
+	 * load(), however this may be used to create an instance if you need to
+	 * use custom SQL queries to load a Model instance
+	 *
+	 * @param row {object} - Row of data returned from the database
+	 *
+	 * @return {Model} The created instance
+	 */
 	Model.createFromRow = function(row){
 		let model = new Model();
 
@@ -142,7 +230,7 @@ function defineModel(table_name, model_fields, options){
 				throw new Error("Cannot construct instance of " + Model.getTableName() +
 				                " since row is missing data field: " + f);
 			}
-			model._fields[f] = row[f];
+			model._fields[f] = Model._mappers[f].fromDb(row[f]);
 		}
 
 		// Ensure model is not marked as dirty
@@ -151,11 +239,15 @@ function defineModel(table_name, model_fields, options){
 		return model;
 	};
 
-	/////////////////////////////////////////////////////////////////////
-	/// \brief Creates an instance of the model from a list of rows
-	/// as returned by database .query
-	/// \return Promise which will evaluate to an array of instances of this Model
-	/////////////////////////////////////////////////////////////////////
+	/**
+	 * Creates an array of Model instances from an array of rows returned from the
+	 * database
+	 *
+	 * @param rows {Object[]} - Array of rows returned from the database
+	 *
+	 * @return {Model[]} Array of created instances. Will be in the same order as
+	 * the rows parameter
+	 */
 	Model.createFromRows = function(rows) {
 		if(rows.rowCount != null && rows.rows != null && rows.fields != null){
 			// This this is the result directly from the database
@@ -169,102 +261,140 @@ function defineModel(table_name, model_fields, options){
 		return result;
 	};
 
-	Model.prototype.isDirty     = function() { return this._dirty; };
+	/**
+	 * Determines if the Model instance is different to the corresponding entry
+	 * in the database, or has not yet been persisted to the database
+	 *
+	 * An instance will automatically be marked as dirty whenever any field
+	 * value is modified. Additionally a new instance will be dirty as it will
+	 * not yet have been persisted to the database
+	 *
+	 * @return {boolean}
+	 */
+	Model.prototype.isDirty = function() { return this._dirty; };
+
+	/**
+	 * Marks the model instance as dirty - IE: differs from the corresponding
+	 * entry in the database.
+	 *
+	 * It is uncommon to need to call this method - modifying any field
+	 * values will automatically set the Model instance as dirty
+	 */
 	Model.prototype.markAsDirty = function() { this._dirty = true; };
-	Model.prototype.getModel    = function() { return Model;       };
 
-	_attachQueryFunctions(Model);
-	_attachJsonSchema    (Model, model_fields);
+	/**
+	 * Retrieves the Model class an Model instance is created from. Typically
+	 * only required to call static methods of the Model if given an instance
+	 *
+	 * @return {Model}
+	 */
+	Model.prototype.getModel = function() { return Model; };
 
-	return Model;
-}
-
-module.exports = {
-	defineModel           : defineModel,
-};
-
-////////////////////////////////////////////////////////////////////////////////
-// Helper functions below
-////////////////////////////////////////////////////////////////////////////////
-
-function _attachQueryFunctions(ModelClass){
 	///////////////////////////////
 	// Load query
-	let stmt_load = "SELECT " + ModelClass.getIdFieldName() + ",";
-	for(let f of ModelClass.getFieldNames()){ stmt_load += f + ","; }
+	let stmt_load = "SELECT " + Model.getIdFieldName() + ",";
+	for(let f of Model.getFieldNames()){ stmt_load += f + ","; }
 	stmt_load = stmt_load.slice(0,-1);
-	stmt_load += " FROM " + ModelClass.getTableName() + " WHERE " + ModelClass.getIdFieldName() + "=?;";
+	stmt_load += " FROM " + Model.getTableName() + " WHERE " + Model.getIdFieldName() + "=?;";
 	//console.info("  Made load stmt: " + stmt_load);
-	ModelClass._queries.load = stmt_load;
+	Model._queries.load = stmt_load;
 
-	/////////////////////////////////////////////////////////////////////
-	/// \brief Loads a specific instance of this model from the data base
-	/// \param dbh A db-connection-promise to use to access the database
-	/// \param id  The id of the instance to load - should be unique
-	/////////////////////////////////////////////////////////////////////
-	ModelClass.load = function(dbh, id){
+	/**
+	 * Loads a specific instance of the Model from the database
+	 * @param dbh {DbConnectionPromise} - Database handle with which to load the
+	 * instance
+	 * @param id - The id of the instance to load - should be unique
+	 * @return {Promise} - Promise which will either resolve to the loaded Model
+	 * instance, or will be rejected if an error occurred during loading
+	 */
+	Model.load = function(dbh, id){
 		return dbh
-			.query(ModelClass._queries.load, [id])
+			.query(Model._queries.load, [id])
 			.then((res) => {
 				if(res.rows.length !== 1){
-					throw ("Failed to load " + id + " from " + ModelClass.getTableName()
+					throw ("Failed to load " + id + " from " + Model.getTableName()
 					       + " since result contained " + res.rows.length + " entries");
 				}
-				return ModelClass.createFromRow(res.rows[0]);
+				return Model.createFromRow(res.rows[0]);
 			});
 	};
 
 	///////////////////////////////
 	// Find query
-	var stmt_find = "SELECT " + ModelClass.getIdFieldName() + ",";
-	for(let f of ModelClass.getFieldNames()){ stmt_find += f + ","; }
+	var stmt_find = "SELECT " + Model.getIdFieldName() + ",";
+	for(let f of Model.getFieldNames()){ stmt_find += f + ","; }
 	stmt_find = stmt_find.slice(0,-1);
-	stmt_find += " FROM " + ModelClass.getTableName() + " WHERE ";
+	stmt_find += " FROM " + Model.getTableName() + " WHERE ";
 	//console.info("  Made find stmt: " + stmt_find);
-	ModelClass._queries.find_prefix = stmt_find;
+	Model._queries.find_prefix = stmt_find;
 
-	ModelClass.find = function(dbh, where_clause, params){
-		//:TODO: escape the where clause?
-		var query = ModelClass._queries.find_prefix + where_clause;
+	/**
+	 * Finds a set of Model instances in the database according to some
+	 * criteria
+	 *
+	 * @param dbh {DbConnectionPromise} - Database handle with which to search
+	 * the database and load the found instances
+	 * @param where_clause {string} - SQL snippet which will be used as the WHERE
+	 * clause when loading
+	 * @param params [array] - Array of values to use for the bound parameters
+	 * in the where clause
+	 */
+	Model.find = function(dbh, where_clause, params){
+		var query = Model._queries.find_prefix + where_clause;
 		return dbh
 			.query(query, params)
-			.then(ModelClass.createFromRows.bind(ModelClass));
+			.then(Model.createFromRows.bind(Model));
 	};
 
 	///////////////////////////////
 	// Update query
 	// :TODO: have option for an auto-update modified_at field
-	var stmt_update = "UPDATE " + ModelClass.getTableName() + " SET ";
-	for(let f of ModelClass.getFieldNames()){
+	var stmt_update = "UPDATE " + Model.getTableName() + " SET ";
+	for(let f of Model.getFieldNames()){
 		stmt_update += f + "=?,";
 	}
 	stmt_update = stmt_update.slice(0, -1);
-	stmt_update += " WHERE " + ModelClass.getIdFieldName() + "=?;";
+	stmt_update += " WHERE " + Model.getIdFieldName() + "=?;";
 	//console.info("  Made update stmt: " + stmt_update);
-	ModelClass._queries.update = stmt_update;
+	Model._queries.update = stmt_update;
 
 	///////////////////////////////
 	// Insert query
 	// :TODO: have option for an auto-update created_at field
-	var stmt_insert = "INSERT INTO " + ModelClass.getTableName() + "(";
-	for(let f of ModelClass.getFieldNames()){ stmt_insert += f + ","; }
+	var stmt_insert = "INSERT INTO " + Model.getTableName() + "(";
+	for(let f of Model.getFieldNames()){ stmt_insert += f + ","; }
 	stmt_insert = stmt_insert.slice(0,-1);
 	stmt_insert += ") VALUES (";
-			for(let f in ModelClass.getFieldNames()){ stmt_insert += "?,"; }
+			for(let f in Model.getFieldNames()){ stmt_insert += "?,"; }
 	stmt_insert = stmt_insert.slice(0,-1);
 	stmt_insert += ");";
 	//console.info("  Made insert stmt: " + stmt_insert);
-	ModelClass._queries.insert = stmt_insert;
+	Model._queries.insert = stmt_insert;
 
-	ModelClass.prototype.save = function(dbh){
+	/**
+	 * Saves the instance to the database. If the instance already exists (IE: id
+	 * is not 0) then will update the existing row. Otherwise will insert a new
+	 * row - after which the id field will be filled in (assuming the id field
+	 * in the database is AUTO_INCREMENT
+	 *
+	 * @param dbh {DbConnectionPromise} - Handle to the database
+	 *
+	 * @return {Promise} Promise which will either resolve to the instance after
+	 * is saved, or will be rejected if an error occurs while updating the
+	 * base. This method will only modify the id field (if the instance has not
+	 * yet been saved) and the isDirty state.
+	 * @todo :TODO: implement auto-updated modified_at etc fields -> this will
+	 * also modify that field
+	 */
+	Model.prototype.save = function(dbh){
 		// Do nothing if the instance is not _dirty
 		if(!this._dirty){ return Q(this); }
 
 		// Generate parameters for the save operation
-		let params = ModelClass
+		let params = Model
 		    .getFieldNames()
 		    .map((f) => {
-			    return this._fields[f];
+			    return Model._mappers[f].toDb(this._fields[f]);
 		    });
 
 		////////////////////////////////////////////
@@ -273,7 +403,7 @@ function _attachQueryFunctions(ModelClass){
 		if(this.id === 0){
 			//then this instance has never been saved before, insert new record
 			promise = dbh
-				.query(ModelClass._queries.insert, params)
+				.query(Model._queries.insert, params)
 				.then((res) => {
 					this._fields.id = res.lastInsertId;
 					return this;
@@ -281,7 +411,7 @@ function _attachQueryFunctions(ModelClass){
 		} else {
 			//updaing existing entry requires a WHERE id = ? as well
 			params.push(this.id);
-			promise = dbh.query(ModelClass._queries.update, params);
+			promise = dbh.query(Model._queries.update, params);
 		}
 
 		promise = promise.then((res) => {
@@ -294,28 +424,50 @@ function _attachQueryFunctions(ModelClass){
 
 	///////////////////////////////
 	// Delete query
-	ModelClass._queries.delete = "DELETE FROM " + ModelClass.getTableName() +
-		" WHERE " + ModelClass.getIdFieldName() + "=?;";
+	Model._queries.delete = "DELETE FROM " + Model.getTableName() +
+		" WHERE " + Model.getIdFieldName() + "=?;";
 
-	//console.info("  Made delete statement: " + ModelClass._queries.delete);
+	//console.info("  Made delete statement: " + Model._queries.delete);
 
-	ModelClass.delete = function(dbh, id){
-		return dbh.query(ModelClass._queries.delete, [id]);
+	/**
+	 * Deletes an instance from the database by id.
+	 *
+	 * Due to the underlying database implementation there is no way to
+	 * test if this operation succeeded - IE: if there does not exist an instance
+	 * with the specified row this method will act as a no-op
+	 *
+	 * @param dbh {DbConnectionPromise} - Handle to the database
+	 * @param id - The id of the instance to delete
+	 */
+	Model.delete = function(dbh, id){
+		return dbh.query(Model._queries.delete, [id]);
 	};
 
-	ModelClass.prototype.delete = function(dbh){
+	Model.prototype.delete = function(dbh){
 		// If not yet saved in database, do nothing
 		if(this.id === 0){ return Q(this); }
 
 		return dbh
-			.query(ModelClass._queries.delete, this.id)
+			.query(Model._queries.delete, this.id)
 			.then((instance) => {
 				this._fields.id = 0;
 				this._dirty     = true;
 				return this;
 			});
 	};
+
+	_attachJsonSchema    (Model, model_fields);
+
+	return Model;
 }
+
+module.exports = {
+	defineModel           : defineModel,
+};
+
+////////////////////////////////////////////////////////////////////////////////
+// Helper functions below
+////////////////////////////////////////////////////////////////////////////////
 
 function _convertToDetailedFieldDescriptors(model_fields){
 	return model_fields.map((field) => {
@@ -356,4 +508,27 @@ function _attachJsonSchema(Model, model_fields){
 
 	Model.schema           = schema;
 	Model.prototype.schema = schema;
+};
+
+function _generateMappingFunctions(model_fields){
+	let result = {
+		id : _mapperNoop,
+	};
+	for(let field of model_fields){
+		if(field.type === 'datetime'){
+			result[field.name] = _mapperDate;
+		} else {
+			result[field.name] = _mapperNoop;
+		}
+	}
+	return result;
+}
+
+let _mapperNoop = {
+	fromDb(x){ return x; },
+	toDb  (x){ return x; },
+};
+let _mapperDate = {
+	fromDb(date_str) { return moment(date_str); },
+	toDb  (date    ) { return date.format("YYYY-MM-DD HH:mm:ss.SSSSSS"); },
 };